CONTRIBUTING: Include note about building modules for test suite
Before running the test suite, the console module at the very least
needs to be compiled. This adds a note about building the modules before
running the test suite.

Additionally, I updated the path for the ArkScript test suite (the
correct path is `tests/arkscript/unittests.ark`).

I've also updated the name of the executable to be `arkscript` so users
do not get the deprecation warning when copying/pasting the command.


Former-commit-id: 835cf7e974ecfc03ca9f427abbe0e9230d1046ba