Compiler: Offer suggestion when symbol is undefined
If the compiler encounters a symbol that is undefined, we can now add a
suggestion user the new `offerSuggestion` method on the compiler.

The suggestion is found by calcluating the Levenshtein Distance between
the symbol that was undefined and the defined symbols. If no viable
suggestion is found, then we simply don't offer a suggestion.