Error Context: Add coloring pairs for parentheses and braces
This adds coloring pairs for parentheses, square braces, and curly
braces. In order to keep track of what pairs are currently open, a new
`LineColorContextCounts` struct has been created.

In addition to that new struct, two new functions have been created to
add colors to the error context: `colorizeLine` and `isPairableChar`.

`colorizeLine` returns the line of code with the pairs of
parentheses/square braces/curly braces colored. The function contains a
vector of termcolor functions that colorize the stingstream. The color
used for the pairing is the modulus of the number of open
parentheses/square braces/curly and the number of termcolors in the
vector. This works nicely since adding or removing coloring pairs simply
requires editing the termcolors vector.