Error Context: Fix underline bug in context error
Underlines weren't showing when context errors were printed. This was
due to some incorrect calculations when writing the context error to the
output stream.

When writing the error line to underline the error, there are two steps:
(1.) print the spaces until we're at the start of the underline and (2.)
print the underline.

This commit refactors the underlining in the Error Context by using two
for loops to clearly show these steps. The first for loop prints the
spaces until we're at the starting column for the underline and the
second for loop prints the underline.