Error Context: Fix VS build errors
This was a bit tricky. Since we're adding colors to the error context
output, we've included `termcolor/termcolor.hpp`. The Termcolor header
file includes `windows.h`, which includes `windef.h`. This last header
file includes `min` and `max` macros. So during preprocessing, the
preprocessor was matching the call to `min` when building on windows and
expanding the macro defined in `windef.h`.

To fix this, we could have included `#define NOMINMAX` in the file,
which would have removed the `min` and `max` macros.

The fix included here seemed like a better fit. This just made the type
on `std::min` explicit so that the preprocessor doesn't match `min` and
leaves it alone.


Former-commit-id: fc2346723026959b522a9f46ff5dbaeb131910c4