refacto: integrating the new parser and enhancing a lot of code surrounding it in the compiler side of things
This new parser isn't ready to use yet, but if I continue, this commit will be 50000 lines by itself, while I just need something that can compile basic files and work the same way as before.
I still need to add an import solver, clean a lot of todo/fixme, (re)implement conditional macros, add the parser's tests in the CI, and some more cleanup, but we are getting there!